fixed transaction amount to Float in queries

+++ app/graphql/types/transaction_type.rb
@@ -1,5 +1,5 @@
 module Types
   class TransactionType < Types::BaseObject
-    field :amount, Integer, null: false
+    field :amount, Float, null: false
   end
 end
